==> prep/bio/1_get_ne_iucn_spp.Rmd <==
---
title: 'OHI-Northeast: 1. Get list of IUCN Species maps in the Northeast'
author: "*Compiled on `r date()` by `r Sys.info()['user']`*"
output: 
  html_document:
    code_folding: hide
    toc: true
    toc_depth: 3
    toc_float: yes
    number_sections: false
    theme: cerulean
    highlight: haddock
    includes: 
      in_header: '~/github/ne-prep/src/templates/ohi_hdr.html'
  pdf_document:
    toc: true
---

# Summary

This script takes the OHI Northeast region shapefile and compares it to the global shapefiles we have from IUCN to extract all species maps that fall within our region. A lot of this code dips into the `spp_risk_dists` repo from Casey O'Hara: https://github.com/jafflerbach/spp_risk_dists.
```{r setup, message = F, warning = F}
knitr::opts_chunk$set(fig.width = 6, fig.height = 4, fig.path = 'figs/',
                      echo = TRUE, message = FALSE, warning = FALSE)
source('~/github/ne-prep/src/R/common.R')

library(sf)
library(raster)

#ne region
ne_shp <- rgns_simp
```


We are going to select only those cells in the Northeast and use that list to query the larger species data. First we have to transform our regions shapefile into the same CRS as the cell ids.

```{r}
#these are global cell ids in raster format pulled from the spp_risk_dists dataset.
cells <- raster("~/github/spp_risk_dists/_spatial/cell_id_rast.tif")

#reproject to our CRS
ne_reproj <- st_transform(ne_shp, crs = "+proj=cea +lon_0=0 +lat_ts=45 +x_0=0 +y_0=0 +ellps=WGS84
+units=m +no_defs")
```

Extract the cell ids for our region
```{r}
ne_cells <- extract(cells, ne_reproj) %>%
  unlist()
```

Read in the necessary files to link species to their map's filepath and species info
```{r}
# .csv file that lists all species and the file path to their map
spp_maps <- read_csv('~/github/spp_risk_dists/_data/spp_marine_maps_2018-1.csv',
                     col_types = 'ddciccc')

#file with species information to link to iucn_sid at the end
spp_info <- read_csv("/home/ohara/git-annex/spp_risk_dists/iucn/spp_info_from_api_2018-1.csv")
```

```{r, eval = F}
#grab each taxa's folder filepath
taxa <- spp_maps$dbf_file %>%
    unique() %>%
    str_replace('\\....$', '')

#create an empty list that is the length of all taxa. We are going to fill this list  
taxa_cells_list <- vector('list', length = length(taxa))

#for each taxa, grab the species map (raster) and filter to only keep those cells in the northeast.
for(i in seq_along(taxa)) { ### i <- 5
    taxon <- taxa[i]
    
    spp_ids_in_taxon <- spp_maps %>%
      filter(str_detect(dbf_file, taxon)) %>%
      .$iucn_sid
    cat(sprintf('processing %s spp in %s...\n', length(spp_ids_in_taxon), taxon))
    
    spp_cells <- parallel::mclapply(spp_ids_in_taxon, mc.cores = 32,
                                    FUN = function(x) { ### x <- spp_ids_in_taxon[1]
                                      f <- file.path('/home/ohara/git-annex/spp_risk_dists/spp_rasters',
                                                     sprintf('iucn_sid_%s.csv', x))
                                      if(file.exists(f)) {
                                        y <- read_csv(f, col_types = 'di') %>%
                                          mutate(iucn_sid = x) %>%
                                          select(-presence)  %>%
                                          filter(cell_id %in% ne_cells)
                                      } else {
                                        y <- data.frame(cell_id = NA,
                                                        iucn_sid = x, 
                                                        f = f, error = 'file not found')
                                      }
                                      return(y)
                                    }) %>%
      bind_rows() %>%
      mutate(spp_gp = taxon)
    
    taxa_cells_list[[i]] <- spp_cells
}
  
taxa_cells_df <- taxa_cells_list %>%
    bind_rows()  %>%
    filter(!is.na(cell_id)) %>%
    select(cell_id, iucn_sid, spp_gp) %>%
    select(iucn_sid) %>%
    distinct() %>%
    left_join(spp_info)
```

We don't have any common names for these maps! So we're using `rfishbase` to get common names

```{r, eval = F}
library(rfishbase)

sp <- validate_names(c(taxa_cells_df$sciname))  
fb_list <- species(sp) %>%
  select(Species, FBname)

#add back in common names
out <- taxa_cells_df %>%
  left_join(fb_list, by = c("sciname" = "Species")) %>%
  rename(common = FBname) %>%
  select(iucn_sid, common, sciname, population, category)
```

We are still missing a lot of common names. We can use the `rredlist` R package to access the IUCN API and grab common names.

```{r, eval = F}
library(rredlist)

sci_names <- out %>%
  filter(is.na(common)) %>%
  .$sciname

big_df <- data.frame()
for(i in 1:length(sci_names)){
  sp <- sci_names[i]

  if(sp == "Phalacrocorax auritus"){ #this one threw some error and no $result was returned from rl_search
    comm <- NA
  }else{
  comm <- rl_search(sp)$result$main_common_name
  }
  
  df <- data.frame(sciname = sp,
                   common = comm)
  
  big_df <- rbind(big_df, df)
}

```

Ok we got a lot of them but not all. Now we can try to use the `taxize` R package.
```{r, eval = F}
library("taxize")

sci_names <- big_df %>%
  filter(is.na(common)) %>%
  .$sciname

common_names <- c()
for(i in 1:length(sci_names)){ #i <- 1
  sp <- sci_names[i]
  comm <- sci2comm(sp, db = "itis")
  
common_names <- c(common_names, comm)
}

#turn list into dataframe
df <- do.call(rbind,lapply(common_names,data.frame))
df$sciname <- rownames(df)

df2 <- df %>%
  rename(common = X..i..) %>%
  filter(!is.na(common)) %>%
  mutate(sciname = str_extract(sciname, "[^.]+")) #remove numbers
```

Only 4 of the species had a common name via taxize. We will add these manually and then the remaining species will keep NA for common name.

```{r, eval = F}
sci_to_comm <- big_df %>%
  mutate(common_name = 
           case_when(
              sciname == "Lestrolepis intermedia" ~ "Barracudina antifaz",
              sciname == "Astrangia poculata" ~ "Northern star coral",
              sciname == "Gigantactis vanhoeffeni" ~ "Cosmopolitan whipnose",
              sciname == "Phalacrocorax auritus" ~ "Double-crested cormorant",
              TRUE ~ as.character(common))) %>%
  select(-common)
```


```{r, eval = F}
#save
write_csv(out, "~/github/ne-prep/prep/bio/data/1_iucn_spp_in_ne.csv")
```

Filter the list of IUCN marine maps from Casey's repo to just have the file paths for species in the Northeast
```{r, eval = F}
spp_marine_maps <- read_csv("~/github/spp_risk_dists/_data/spp_marine_maps_2018-1.csv") %>%
  filter(sciname %in% out$sciname,
         subpop %in% c("Atlantic Northwest", "Northwest Atlantic", NA)) #all other subpops that have regions are not in the Northeast so remove here
#save
write.csv(spp_marine_maps, "~/github/ne-prep/prep/bio/data/1_iucn_spp_shp_filepaths.csv")
```
